+++ NEWS.md
@@ -1,9 +1,11 @@
-# SpatDiv 0.1-0.9001
+# SpatDiv 0.1-0.9003
 
 ## New features
 
-* Species distribution S3 methods : `as.AbdVector()`, `as.ProbaVector()` for `wmppp` objects.
+* Species distribution S3 methods : `as.AbdVector()`, `as.ProbaVector()` for `wmppp` objects, factors and character vectors.
 
-* `Tsallis()` and `Diversity()` for `wmppp` objects.
+* `Tsallis()` and `Diversity()` for `wmppp` objects, factors and character vectors.
 
 * `rSpCommunity()`generates a spatialized random community (in a `wmppp` object). In progress: limited to a binomial point process.
+
+* `EntAccum()` and `DivAccum()` calculate entropy and diversity accumulations curves.